requirements.extracted.txt
# Raw extraction from existing module requirement files.

## agri_qa_engine/requirements.txt
openai==2.24.0
python-dotenv==1.2.1

## KnowledgeRetrieval/requirements.txt
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
torch==2.2.2
transformers==4.38.2
faiss-cpu==1.7.4
nltk==3.8.1
networkx==3.2.1
flask==2.1.3
pydot==1.4.2
matplotlib==3.8.4

## KnowledgeGraph/requirements.txt
scrapy==2.11.1
requests==2.31.0
python-docx==1.1.0
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow==10.2.0
beautifulsoup4==4.12.3
lxml==4.9.4
jieba==0.42.1
scikit-learn==1.3.2
spacy==3.7.2
transformers==4.39.3
torch==2.2.1
snownlp==0.12.3
py2neo==2021.2.3
neo4j==5.17.0
networkx==3.2.1
schedule==1.2.1
pydantic==2.6.4
pydantic-settings==2.2.1
tqdm==4.66.2
